audio_path: Switch to use real time state detection
Do not use sharedpreferences on this menu, sometimes when user or system switch it by their algorithm then M.A.O can't detect real state for that switch is in active or non active condition.

So switch to use file dump debugging to always stay at real time condition when layout get refreshed or re-start activity.