audio_stats: Check audio low latency mode at first
Check state if device use audio low latency mode or not, check this at first can make M.A.O detect it correctly what to do with ULL enable or not.

But sometimes on new syberia build, DIRECT flags can't be closed, so M.A.O will read that DIRECT flags although deep_buffer that's actual flags on use.

Will try to solve it later