audio_conf: Use proper handling for unavaible feature
Disabling cardview when kernel feature is not available in the device isn't a good way, User will not know what is that and why is only empty.

To avoid this, better to hide only switch and show another textview to tell user if that feature is not exist in their kernel.